Add configuration diff
Our configuration file changes pretty often, e.g. when introducing new configuration values.
We require the users to have an up to date configuration file.
In the usual workflow, it is hard to see if a value or key changed in the template.
This commit automated the comparision.
A developer can now see directly when a value is changed.