Assign branch and depth to Operator
Ensure start branch and depth have proper formats. Set history depth to 1 if only latest commit is desired and otherwise to 0 (in case of complete history). After session creation and cloning of git repository, delete attribute branch and depth from dictionary body because they are only needed for the cloning of the git model.