For readonly sesssions: Allow to choose a specific branch/tag (or all) and the depth (latest commit or complete commit history) after selection of project repository
After selection of a repository for a readonly session, two selection fields are computed (implemented by the attribute chosenRepository) where to choose the branch/tag and the depth. A FormControl is implemented for these variables and default are set such that the cloning of the git model is as fast as possible, i.e. branch is set to main branch and depth to latest commit. Add also after repository selection variable 'All' to available branches for cloning the complete git model.