fix: request of persistent sessions
Previously, it was not possible to click on the request button when the workspace was set to persisten session. The reason was that the conditions of repositoryFormGroup couldn't fulfilled since no reference or depth couldn't be chosen. Therefore, the FormControls, reference and historyDepth, were taken out and added to an additional FormGroup, referenceDepthFormGroup (which is used just for read only sessions). So, the request button is only disabled when repositoryFormGroup is not valid and in case of read only sessions, referenceDepthFormGroup has also to be valid