refactor: Change storing and retrieving method
Instead of simply storing the instance within the service,
you now need to subscribe to an observable that will
always provide the last loaded instance. By doing this,
components using the service can assume that they
always have the most recent instance
Also, one are now loads an instance with a post because
the credentials are placed inside the http body.