refactor: Move modelsources into projects folder
This moves the Git and T4C model sources to the project
folder, as they are a sub-resource of projects and their
models. As a result, some frontend routes and some Python
imports had to be updated. Also, some code improvements
were made, get_project was renamed to get_project_by_name
and some Fast Api router tags were adjusted.