feat: Add validation & make primary functionality
This provides backend validation for retrieving and updating
git models. The validation checks whether a user is allowedto access or update a git model, whether the git
model actually exists for the specified project and capella
model, and whether the provided git model meets the
requirements. In addition, the ability to make a non-primary
git model a primary one is added in the frontend.