refactor: Add verification by injection (projects)
This changes all uses of the project name to the project slug
(for the standard /projects routes), since the name should
no longer be used (except to create a project). In addition,
the convert_project function is removed and integrated into
the project itself as a validator (this eliminates the need to
call convert_project explicitly).