feat: Include asynchronous prefix validator
This revises the previous validator for the result url. First, it no
longer checks if the base or input url is valid, since the asynchronous
validator is only executed if both are valid. In addition, we changed
the use of the updateResultUrl method with the execution of the new
asynchronous validator, as it should always be called when there is a
change in the result url to ensure that the disabled fields are enabled.