feat: Adjust kubernetes operator
First, the definition of Kubernetes objects is changed via a dictionary
and now uses Kubernetes client. Also, promtail is added to the
session deployments (and the configmap used by promtail) and some
context information is added as labels for the session deployments.
In addition, some minor improvement to the kubernetes operator
has also been added.