fix!: Remove infinite loop during authentication
When the access token is expired, the client sends a request to the backend to renew it with the renew token.
However, if the renew token is expired as well, the backend sends an unauthenticated response.
That has changed, before it was an internal server error.
The frontend reacts to the "Unauthenticated" by sending a request to renew the access token again...
Which leads to an infinite loop and our users are basically spaming the backend with requests.