feat: Print errors in `/healthcheck`
This commit also prints all messages with LEVEL >= INFO for the healthcheck routes.
This is particually interesting because we're not able to filter for extra arguments
in the logging filter as `logging.LogRecord` doesn't contain extra attributes.
Therefore, we're just filtering for the message. The filtering is more specific now,
it also checks for the `path=` prefix.