feat: Add job layer for pipelines
Currently, it's only possible to get the logs of the last pipeline.
Even this wasn't working properly, sometimes logs of old pipelines were displayed.
In general, the logs were also not persistent and there was no job status.

This commit adds another abstraction layer for manually triggered jobs.
For each pipeline, we'll show a list of jobs. Each job has a triggerer,
a start time, a job state and logs/events. Job logs and events are stored in Grafana Loki
and are persistent.

Nightly jobs are not part of this commit and are not shown on the job list.

In addition, we'll now also detect unknown jobs in the cluster and delete those.
This can be seen as garbage cleaning if the deletion of jobs fails, e.g. due to a server failure.