feat: Adjust models for alchemy 2.0 compatibility
This commit incorporates significant changes to our SQLAlchemy
models to align with the new SQLAlchemy 2.0 format. The most
notable change is the introduction of `Mapped[<type>]` for
annotating field types, which enables automatic detection of
database column types.
A primary focus of this PR was the evaluation and modification
of each field's nullability status based on the new typing
provided by SQLAlchemy.
In a follow-up PR, we can continue to enhance our database classes
by turning our database classes into
[dataclasses](https://docs.sqlalchemy.org/en/20/orm/dataclasses.html).