feat: Adjust CRUDs for alchemy 2.0 compatibility
This commit introduces substantial revisions to our CRUD
operations. A primary change includes transitioning from the legacy
`db.query(...)` method to the new `db.execute(...)` standard
offered by SQLAlchemy 2.0.
The PR also aims to improve the consistency of our CRUD functions
through the following modifications:
- The `db: Session` object is now uniformly the first argument
across all functions.
- All 'get' CRUD functions will return the requested object or
`None`. If an object is not found and an error should be raised,
it is advised to either use a `_or_raise` CRUD function that raises
an SQL exception or to handle the situation outside of the CRUD
file, possibly by throwing an HTTP Exception.