refactor: Import adjustments to correct style
Up to now, many of our imports were inconsistent because they did
not conform to the Google Import Style Guide we use. This PR now
fixes that for several specific import types, as well as some import
improvements on top of that. The following imports have been
consistently adjusted and should no longer show up in the code:
- `from sqlalchemy.orm import Session`
- `from sqlalchemy import select, ...`
- `... import `get_db`
- `from fastapi import ApiRouter, Depends, ...`

However, due to the amount of non-complient imports, not all
issues were resolved with this PR. This should either be done in a
follow-up PR or better, they should be updated when touched or used.