fix: Failing backups for Capella >= 6.0.0 without HTTP port
Our backup image defauls to the HTTP connection method
for all Capella versions >= 6.0.0.
However, the Collaboration Manager is not passing the
HTTP_PORT, HTTP_USERNAME and HTTP_PASSWORD variables
when the http port in not set.

If the HTTP port is not set, we should fallback to the
telnet connection method.
Therefore, the CONNECTION_TYPE variable is passed explicitly.