feat: Manual focusing of sessions in session viewer
Security features of modern browsers make it hard to detect the focused iframe.
This commit indicates which session is currently focused.
The focused session has access to the keyboard, the other session doesn't.
We're trying to auto-detect which session is focused, but it doesn't always work.
In this case, the user can manually focus the session by clicking on the title bar.