feat: Restrict parallel active sessions
In general, we don't allow parallel sessions.
This limitation is valid for Eclipse-based sesssions as
Eclipse locks workspaces. Juypter itself is independent and
can run in parallel to Eclipse-based sessions.

Jupyter & Eclipse-based tools in parallel and a requirement for the
session viewer. Therefore, this serves just as workaround
until we have proper handling for workspace management and workspace
locking.