docs: Move release notes from docs to Github release notes
The initial idea behind release notes in the documentation
was to update them during PRs. However, this has not worked
as expected, leading to outdated release notes.

I've moved all existing release notes to the already existing
Github releases (in the "Notable changes") section.

Also, I added the videos from PRs to the release notes.

The release notes page in the documentation now points to
the individual repositories.