feat: Create project-level file share for Jupyter
A file-share is created once a Jupyter model is added to a project.
Per default, the size of the volume is 2GB.

In this commit, the logic for the creation of volumes was moved
out of the operator and to the corresponding modules (sessions / toolmodels).
The operator should not contain any logic.

The volume is not yet mounted, this will be added in a later commit.
Mounting requires more severe changes to the operator.