feat: Introduce session hooks
The sessions routes and the kubernetes operator grew a lot in the past.
These two files can be hard to understand and code quality was rather bad.

A lot of tool-specific logic happened in the two files, e.g.,
the creation of a T4C session token.
This logic doesn't belong to the operator or the sessions module,
the context is tool-specific, but it has a strong connection to the
sessions lifecycle.

The solution is session-hooks. For the beginning, we support the following hooks:
- configuration_hook (Configuration for the session)
- post_session_creation_hook (Hook executed after session creation)
- post_session_termination_hook (Hook executed after session termination)

The T4C, pure::variants and Jupyter logic was moved to session-hooks.