feat: Add volumes to configuration session hook
Volumes can be defined in the configuration hook.
The volumes are then mounted into the session containers.

For the beginning, the persistent workspace is moved to
a new PersistentWorkspaceHook. Therefore, the
`persistent_workspace_claim_name` parameter could be removed
from the session creation call chain.

The feature can be extended to also support secrets
(for pure::variants license mounting) and project-level file-shares.