feat: Mount project-level file-shares into Jupyter sessions
This is a follow-up of https://github.com/DSD-DBS/capella-collab-manager/pull/894.

In #894, project-wide Jupyter file-shared were added, but not mounted
to sessions. This missing bit is added in this commit.

The shares are mounted to `/shared/<project_slug>/<toolmodel_slug>`.
To access the shares directly, a symlink from `/workspace/notebooks/shared`
to `/shared` is required in the session container. This was added in
https://github.com/DSD-DBS/capella-dockerimages/pull/187.

We do consider the access permission for users in projects:

- All users can see the file-shares from internal projects in
  a read-only mode.
- Administrators can access all file-shares in a read/write-mode
- Project members of private projects have access according to
  their permission (read-only or read/write).
- If a user has no access to a project, file-shares of the project
  are not visible for the user.

To prevent data-loss, we forbid writing files to `/shared` and
`/shared/*`. These paths can not be mapped to a file-share.