fix(front): Correctly reload restrictions after update
This commit fixes the model restrictions update in the frontend.
The `loadModelBySlug` function was called with the wrong argument order,
so when trying to refresh the model the project and model slug
where swapped. In addition, this commit also contains a few
improvements to removed the usage of `!` and `JSON.stringify`.