feat: Improve diagram cache margins on mobile phones
The diagram overview and diagram viewer makes now full use
of the available space on mobile phones.
As the previous way to close the diagram by just pressing next to it
it not available anymore (there are no margins),
there is a new sticky close button on mobile phones.

The UI on devices with wider screens (>=1024px) doesn't change.