feat: Make Guacamole database persistent in development mode
When running the environment with target: local, the Guacamole database
was not persistent. The database had to be reinitialized after each restart.

This can be annoying, also in local development environments.
The specific Guacamole database handling for local environment
is removed in this commit.

This will improve the local stability, see #800.