feat: Add support for wildcard hosts
In some environments, e.g. local unmanaged single-node clusters, it might
make sense to serve the application on another host.

This feature was requested in #745.

Kubernetes ingress has support for wildcard hosts. I can be enabled
by just setting the host to `null` in the ingress definition.

Also, the commit fixes an issue with reverse proxies, where the application
didn't serve anymore because the internal host didn't match the
one configured in the `values.yml`.