feat: Ensure that PostgreSQL database uses 'UTC' timezone
The application stores the datatime objects in a timezone naive format.
Therefore, we store them as UTC and also expect to get UTC as response.

We expect the database to run in UTC, it should not perform any additional
datetime or timezone transformation. The transformation is done in the
application.

However, some databases run in different timezones. In this case, we
ensure that UTC is used for our transactions.