refactor: Use dataclasses for all SQLAlchemy database classes
With dataclasses, a `__init__` function is created for all database
classes. This adds autocomplete and code recommendations and type
checkers like mypy can check the passed types.

Since the `__hash__` method was removed during this change, the intersection
method to get common projects doesn't work anymore. I changed it to a proper
database join, which is faster.

In addition, many type annotations were added. Some mypy and pylint errors
and warnings were fixed in the tests.