feat: Add connection method and custom variables for tools
Tools can now support different connection methods, for example
Guacamole, http (/xpra). The available options can be configured in
the tools YAML editor.

If configured properly, the connection methods are displayed during
persistent session creation and are selectable by the users.

To reduce complexity in the backend, the Guacamole configuration
is moved to a session hook now.

A new session hook `fetch_session_hook` has been introduced.
It is executed during fetching of each session (after creation /
in the GET route, not in the GET ALL routes).

Session environment variables can be defined per tool now in a flexible way.
The existing Jupyter environment variables were migrated to the new format.

All references to Guacamole except the Guacamole status in the Monitoring view
are removed from the frontend. All Jupyter specific code is removed
from the frontend and was replaced with general web-based implementation.

Most of the Guacamole and Jupyter code was removed from the backend.