fix(frontend): Remove trailing `/` from backend API requests
A trailing slash leads to an unnecessary redirect, triggered by the backend.
In addition, it's harder readable when a `/` is part of a base URL, which is then used
in template literals, e.g. `${baseURL}/${resource}` is better to read than
`${baseURL}${resource}`.

The redirect behaviour lead to occasional request failues and some parts
of the frontend stopped working when running behind a reverse proxy with TLS edge termination.