feat: Move pure::variants secret mounting to session hook
To achieve this, a new volume type "SecretReferenceVolume" was added.

The new volume can be used to mount existing secrets to a container.
It's required for mounting the pure::variants license file, which was
part of the k8s.py before. In this commit, it's moved to the p::v session hook.

In addition, the commit contains many test improvements for pure::variants