docs: Shorten URLs with duplicate names in path
We use nesting in our documentation files to store screenshots
in the same context with the markdown file.

This usually adds another nesting layer in the URL, e.g.
the path `getting_started/getting_started.md` was rendered to
`getting_started/getting_started`.

Instead, this could be simplified by using a file with the name `index.md`:
The path `getting_started/index.md` renders to `getting_started`.

This looks cleaner, shortens the URLs and removes duplication.