fix: Correctly handle disabling version selection
This moves the logic of enabling version selection from an attribute
to the actual logic. First of all, this removes the warning that angular
always showed in the console about how we handled disabling
the selection here. Second, once a tool is selected, there is no way
to go back to the state where the version dropdown should be disabled,
so just handling the enable case in the logic is quite simple.
Finally, the current approach was not working properly because
even without a tool selected, the version selector seemed to be
enabled, i.e. we were even able to interact with it.