fix(front): Update only new or modified sessions
This first introduces the new angular control flow in this component.
By doing so, we can easily add `track session.id` to the session loop.
By tracking sessions by their id, only newly created or modified
sessions are reloaded in the DOM. This not only improves performance
by reducing the number of reloads, but also allows us to properly
select parts of the session information, since previously each reload
removed the selection. Finally, we also cleaned up the unused
`sessions` variable and converted most of the styling to tailwind.