refactor: Apply control flow to diagram cache
This changes the previous usage of `ngIf` and `ngFor` to the new
angular control flow introduced in angular v17. This not only improves
code readability, but also allows better alignment and combination
of conditions. To test this, we introduced diagram cache stories
in #1549, which can be compared to the stories in this PR.