fix: Trust proxy headers to avoid https -> http redirects
When FastAPI sent a temporary redirect, it didn't respect the `X-Forwarded-Proto`
header of the proxy.

Therefore, it did default to `http` and sent an http URL as response.

In this commit we tell Uvicorn to trust all IPs. Since the backend is firewalled
and only reachable behind the Kubernetes Ingress / proxy, this is safe.