feat: Improve styling of footer & header
The position of the footer was below the screen.
The scrollbar was always visible and users had to scroll to see
the full footer.

This commits shows the footer sticked to the bottom when the page size
is smaller than the screen.

In addition, the legal component was moved to the footer component.
TailwindCSS was introduced in those components.

ngIf was replaced with the new Angular control flow.

The footer component is now responsive and text alignment is now consistent
on mobile devices (before, some parts were centered and other parts were left-aligned).

Stories for the Version-, Header- and FooterComponents were added.

The header uses flat buttons now as they look slightly cleaner.

The "Open in GitHub" button was removed since people have used it to report
internal issues. This can lead to potential exposure of confidential content.

The "Capella Collaboration Manager" text was removed on mobile devices to
save space on the small screens.