feat: Rework authentication logic
This commit serves two main purposes. First, it merges the code
for the two authentication providers, azure and oauth. To do this,
we no longer rely on the azure-specific MSAL library, but use a more
general library that we already use for non-azure providers.
Second, we change our authentication scheme from a
bearer token provided in the header and stored in a frontend service
to secure cookies. This eliminates an XSS attack vector as
the tokens are no longer accessible on the client side.