docs: Document how to setup Keycloak as CCM IdP
This documents how to use Keycloak as an identity provider
for the capella Collaboration Manager (CCM). This includes how to
properly configure a CCM client in Keycloak and how to set up
the configuration in CCM to use that client. However, it is important
to note that the documentation does not cover how to properly
configure and securely use Keycloak.