feat: Introduce idp identifier and email columns
This commit primarily adds two columns to the user table,
the idp identifier and the email. The idp identifier is used to map
the idp user to the local user. Therefore, it is critical that the idp
claim used for this is unique per user. The following breaking changes
are introduced to the values.yaml file: The `jwt.usernameClaim' is
now moved to `claimMapping.username' and specifies the identity token
claim used for the username column. The `claimMapping.idpIdentifier'
is added, which specifies the identity token claim used for the
new idp identifier column and must be unique within the idp.
The `claimMapping.email` is added, which specifies the identity token
claim used for the new email column. The breaking changes are
detailed in the PR description and in the release notes.