refactor: Wait for database container to start
While running the tests, I kept getting an `IndexError' for
retrieving the database container's external port for the init
database test. This change now properly waits for the container to run.
If the container is already running, the behavior is exactly the same
as before, i.e. we only wait if the container is not already running.