fix: Remove `WWW-Authenticate` header from responses
The HTTP specification recommends to send a `WWW-Authenticate` header with
status code 401. We've used that to indicate which authentication methods
are available for clients, in our case Basic and Cookie.

While this works fine on Chrome, Safari starts prompting for basic authentication
on its own. I haven't found a way to suppress, so we'll have to remove the
`WWW-Authenticate` header to avoid this behaviour.