fix: Add CORS Headers to HTTP error 500 responses
Internal server error respones result in no CORS ALLOW_ORIGIN headers
added to the response. This leads to wrong handling by the frontend.
When there are no CORS headers, status code was 0, which was treated as
"Backend not reachable".

There was already handling to add CORS headers, but it didn't add the
ALLOW_ORIGIN header, which we add manually now.