feat: Let session proxy handle timeouts
The default timeout for nginx is 60 seconds and for haproxy it's 30 seconds.
Our session proxy should terminate requests after 60 seconds with a custom
error page, but the Ingress nginx / HAProxy were faster.

This sets the timeout to 70s, independently from nginx / HAProxy.
Since we want to have custom annotations for sessions and the rest of the
application, this commit splits the ingress into two.