fix: Don't loop-redirect to `/auth` or `logout`
If the redirect is set to one of those URLs, just redirect
to the start page. Otherwise this will create a redirect loop since the user
is logged out after login.

Also improves security by removing the query params from the callback URL while
the token request is in progress.