fix: Fix an issue where feedback that shared contact info was not sent
Feedback emails are sent in a background task to decrease response latency. Unfortunately, this
means that the SQLAlchemy session is already closed and none of the fields of DatabaseUser can be
resolved. This bug was not found because the error messages are swallowed by the background task.
The bug is fixed by turning the DatabaseUser object into a normal User object before passing it to
the background task.