fix: Increase connection pool to 20 concurrent database connections
Due to the asynchronous nature of FastAPI and the number of requests,
it can happen that we have to handle more than the default of 5 requests
at the same time.

In addition, reduce the pool timeout to 5. The goal is to fail requests
if they can't get a database session within 5 seconds instead of spawning
a queue with wait times of 30 seconds, which can lead to a blocking backend.