chore: Enable pylint in tests folder
Enable pylint in the tests folder. The pylint-pytest plugin is used to prevent false-positives for
fixtures. There are still some false-positives for unused arguments in methods and classes that we
mock, so corresponding comments to ignore these errors are added. This was also used to fix some
valid issues that pylint found (like duplicate and unnecessary imports).

Closes #544