feat: Add fine-grained permission system using scopes
Instead of the inflexible roles and project roles,
introduce a fine-grained permission system.
During creation of personal access tokens,
the user can select the permissions that the token should have.
This limits the previously unrestricted access of personal access tokens.

For OpenID / frontend authentication,
the role concept is still in place and will be expended
with customizable roles in the future.
The limitation that administrators can't be added to projects
has been lifted; this is a preparation for custom roles in the future.

This PR is a basis for the implementation of AD groups,
which can be mapped to the new fine-grained permission format.

The PR also changes the way exceptions are shown in the API documentation.
Examples are now generated per exception,
making it easier to reference them in route responses.
Exceptions can now also be referenced from dependencies.