feat: Add fine-grained permission system using scopes
Instead of the unflexible roles and project roles, introduce a
fine-grained permission system.

For OpenID / frontend authentication, the role concept is
still in place and will be expended with customizable roles in the future.

During creation of personal access tokens, the user can select
the permissions that the token should have. This replaces the full-access
of personal access tokens. Existing access tokens will receive all
permissions that the user has access to during migration.

The permission system will be matched the current role and the behaviour
will not change; it's not a breaking change.

This is also the basis for mapping of OAuth scopes.