perf: Replace custom timezone validation with official one
Our custom validator used zoneinfo to load the list of timezones every time. The python docs say
this is bad for performance, and from some profiling it looks like this was taking a decent amount
of time. The official validator seems to cache the list and performs a lot better.