build: Add compatility with `capella-dockerimages>=3.0.0`
CDI had some breaking changes in v3.0.0 and this is the adaption to it.
As part of it, `CAPELLA_VERSIONS` and `T4C_CLIENT_VERSIONS` were replaced
with `CAPELLA_VERSION`.

The default Capella version was changed to Capella 7.0.0.

The Makefile is no longer compatible with `capella-dockerimages<3.0.0`.
This does explicitly NOT affect the CCM compatility with older images.