feat: Make slugs of projects and tool models immutable
Before, the slug was auto-updated when the name of a project or
tool model was changed. However, the relevance of the slug grew over time
and a change can break many parts, in particual:

- The slugs are used for the file-paths of the training provisioning
- The slugs are used as unique identifier for the API; existing script
  may break after an update.

It's indended to add another feature to update the slug individually in the future.
Until then, a new project can be created and all models can be moved to the new project.