refactor: Use Valibot instead of Zod to decrease bundle size
My original plan was to migrate this project to Zod v4 mini (which has
a similarly small bundle size), but as this requires certain tsconfig
options that don't work with Angular, I decided to replace Zod with
Valibot. This also fixes a bug where an invalid session history would be
permanently corrupted instead of automatically resetting itself.